Move button group to machines/components
