+++ src/app/machines/components/MachineActionButtonGroup/index.scss
